+++ src/app/event-form/event-form.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Optional, Self } from '@angular/core';
 import { NgControl, Validators } from '@angular/forms';
-import { typedFormGroup, typedFormControl, TypedControlsIn } from '../shared/forms-util';
+import { typedFormGroup, typedFormControl, TypedControlsIn } from 'forms';
 import { EventForm, eventDefault } from './event-form.model';
 import { ControlValueAccessorConnector } from '../shared/control-value-accessor-connector';
